Create Date-time calculator
I did a lesson on datetime module and I then decided to go off on my own tangent.  I started with some calculations of days alive, months alive etc. I then proceeded to try and import the current day and make the pc display the current day as a word.  I tried making the days of the week into a dictionary and i created a key-value pair to try and index into the dictionary and compare that with the current day.  I am not getting it right for some reason and I think it has to do with me comparing two values that are not being seen as equal.  I think I need to make both values an integer or a string in order to get the comparrison to work.  Basically I am running a loop to find a match between the current day and the day in the week in the dictionary that matches it (or at least that is what I am trying to do).