Create Mandala drawing turtle
This was right at the beginning of my programming journey when I learned about turtle for the first time.  I learned hoe to create a square and how to run a loop.  I took this knowledge and made the turtle draw a regular mandala and then I tweaked the numbers to make it draw a different and more complex mandala.  Pretty simple and straight forward but very satisfying to see my first attempt at a bit of coding come together.